CHANGE: Split Input implementation into multiple files.
