+++ Packages/com.unity.input/Runtime/Input.cs
@@ -1,10 +1,7 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using UnityEngine.InputSystem;
-using UnityEngine.InputSystem.EnhancedTouch;
 using UnityEngine.InputSystem.LowLevel;
-using UnityEngine.InputSystem.Utilities;
 using UnityEngineInternal;
 
 // GOAL: Show how you can load an existing Unity project using legacy input, install com.unity.inputsystem, and play it as is while actually not using the old input system at all
@@ -38,233 +35,6 @@ namespace UnityEngine
         FaceDown = 6
     }
 
-    public struct AccelerationEvent : IEquatable<AccelerationEvent>
-    {
-        internal float x, y, z;
-        internal float m_TimeDelta;
-
-        public Vector3 acceleration => new Vector3(x, y, z);
-        public float deltaTime => m_TimeDelta;
-
-        public bool Equals(AccelerationEvent other)
-        {
-            return Mathf.Approximately(x, other.x) && Mathf.Approximately(y, other.y) && Mathf.Approximately(z, other.z) &&
-                Mathf.Approximately(deltaTime, other.deltaTime);
-        }
-
-        public override bool Equals(object obj)
-        {
-            return obj is AccelerationEvent other && Equals(other);
-        }
-
-        public override int GetHashCode()
-        {
-            return HashCode.Combine(x, y, z, m_TimeDelta);
-        }
-
-        public static bool operator==(AccelerationEvent left, AccelerationEvent right)
-        {
-            return left.Equals(right);
-        }
-
-        public static bool operator!=(AccelerationEvent left, AccelerationEvent right)
-        {
-            return !left.Equals(right);
-        }
-    }
-
-    public struct LocationInfo
-    {
-        internal double m_Timestamp;
-        internal float m_Latitude;
-        internal float m_Longitude;
-        internal float m_Altitude;
-        internal float m_HorizontalAccuracy;
-        internal float m_VerticalAccuracy;
-
-        public float latitude => m_Latitude;
-        public float longitude => m_Longitude;
-        public float altitude => m_Altitude;
-        public float horizontalAccuracy => m_HorizontalAccuracy;
-        public float verticalAccuracy => m_VerticalAccuracy;
-        public double timestamp => m_Timestamp;
-    }
-
-    public enum LocationServiceStatus
-    {
-        Stopped = 0,
-        Initializing = 1,
-        Running = 2,
-        Failed = 3
-    }
-
-    public class Compass
-    {
-        public float magneticHeading => InputRuntime.s_Instance.lastHeading.magneticHeading;
-        public float trueHeading => InputRuntime.s_Instance.lastHeading.trueHeading;
-        public float headingAccuracy => InputRuntime.s_Instance.lastHeading.headingAccuracy;
-        public Vector3 rawVector => InputRuntime.s_Instance.lastHeading.raw;
-        public double timestamp => InputRuntime.s_Instance.lastHeading.timestamp;
-        public bool enabled
-        {
-            get => InputRuntime.s_Instance.headingUpdatesEnabled;
-            set => InputRuntime.s_Instance.headingUpdatesEnabled = value;
-        }
-
-        internal struct Heading
-        {
-            public float magneticHeading;
-            public float trueHeading;
-            public float headingAccuracy;
-            public Vector3 raw;
-            public double timestamp;
-        }
-    }
-
-    public class Gyroscope
-    {
-        public Vector3 rotationRate => gyroscopeSensor?.angularVelocity.ReadValue() ?? default;
-        // So apparently, iOS is the only platform where rotationRateUnbiased is a thing. Other platforms seem to just
-        // return rotationRate here. In the current Sensor API of InputSystem, we are not surfacing this information.
-        // For now, go and just return rotationRate on all platforms.
-        public Vector3 rotationRateUnbiased => rotationRate;
-        public Vector3 gravity => gravitySensor?.gravity.ReadValue() ?? default;
-        public Vector3 userAcceleration => accelerationSensor?.acceleration.ReadValue() ?? default;
-        public Quaternion attitude => attitudeSensor?.attitude.ReadValue() ?? default;
-        public bool enabled
-        {
-            ////REVIEW: Should this return true if *any* is enabled? Right now, failing to turn on any one single sensor seems makes it seem like the entire gyro failed to turn on.
-            get => IsEnabled(gyroscopeSensor) && IsEnabled(accelerationSensor) && IsEnabled(gravitySensor) && IsEnabled(attitudeSensor);
-            set
-            {
-                SetEnabled(gyroscopeSensor, value);
-                SetEnabled(accelerationSensor, value);
-                SetEnabled(gravitySensor, value);
-                SetEnabled(attitudeSensor, value);
-            }
-        }
-        public float updateInterval
-        {
-            ////REVIEW: Not clear which to pick here; current code goes for min frequency of any of the actual sensors.
-            get => Mathf.Min(GetFrequency(gyroscopeSensor),
-                Mathf.Min(GetFrequency(accelerationSensor), Mathf.Min(GetFrequency(gravitySensor), GetFrequency(attitudeSensor))));
-            set
-            {
-                SetFrequency(gyroscopeSensor, value);
-                SetFrequency(accelerationSensor, value);
-                SetFrequency(gravitySensor, value);
-                SetFrequency(attitudeSensor, value);
-            }
-        }
-
-        // The UnityEngine.Input gyroscope is actually an aggregation
-        // of multiple sensor types. So, we hold on to multiple devices
-        // here.
-        private InputSystem.Gyroscope m_Gyro;
-        private LinearAccelerationSensor m_Accel;
-        private GravitySensor m_Gravity;
-        private AttitudeSensor m_Attitude;
-
-        private InputSystem.Gyroscope gyroscopeSensor
-        {
-            get
-            {
-                if (m_Gyro == null || !m_Gyro.added)
-                    m_Gyro = InputSystem.Gyroscope.current;
-                return m_Gyro;
-            }
-        }
-
-        private LinearAccelerationSensor accelerationSensor
-        {
-            get
-            {
-                if (m_Accel == null || !m_Accel.added)
-                    m_Accel = LinearAccelerationSensor.current;
-                return m_Accel;
-            }
-        }
-
-        private GravitySensor gravitySensor
-        {
-            get
-            {
-                if (m_Gravity == null || !m_Gravity.added)
-                    m_Gravity = GravitySensor.current;
-                return m_Gravity;
-            }
-        }
-
-        private AttitudeSensor attitudeSensor
-        {
-            get
-            {
-                if (m_Attitude == null || !m_Attitude.added)
-                    m_Attitude = AttitudeSensor.current;
-                return m_Attitude;
-            }
-        }
-
-        private bool IsEnabled(InputDevice device)
-        {
-            return device != null && device.enabled;
-        }
-
-        private void SetEnabled(InputDevice device, bool enabled)
-        {
-            if (device == null)
-                return;
-
-            if (enabled)
-                InputSystem.InputSystem.EnableDevice(device);
-            else
-                InputSystem.InputSystem.DisableDevice(device);
-        }
-
-        private float GetFrequency(Sensor device)
-        {
-            if (device == null)
-                return default;
-
-            return device.samplingFrequency;
-        }
-
-        private void SetFrequency(Sensor device, float frequency)
-        {
-            if (device == null)
-                return;
-
-            device.samplingFrequency = frequency;
-        }
-    }
-
-    public class LocationService
-    {
-        public bool isEnabledByUser => InputRuntime.s_Instance.isLocationServiceEnabledByUser;
-        public LocationServiceStatus status => InputRuntime.s_Instance.locationServiceStatus;
-        public LocationInfo lastData => InputRuntime.s_Instance.lastLocation;
-
-        public void Start(float desiredAccuracyInMeters, float updateDistanceInMeters)
-        {
-            InputRuntime.s_Instance.StartUpdatingLocation(desiredAccuracyInMeters, updateDistanceInMeters);
-        }
-
-        public void Start(float desiredAccuracyInMeters)
-        {
-            Start(desiredAccuracyInMeters, 10f);
-        }
-
-        public void Start()
-        {
-            Start(10f, 10f);
-        }
-
-        public void Stop()
-        {
-            InputRuntime.s_Instance.StopUpdatingLocation();
-        }
-    }
-
     /// <summary>
     /// Interface for accessing and managing input.
     /// </summary>
@@ -273,17 +43,7 @@ namespace UnityEngine
     /// </remarks>
     public static partial class Input // Partial to allow us to inject project-specific generated code into the API.
     {
-        public static bool mousePresent => Mouse.current != null;
-        public static Vector2 mousePosition => Pointer.current?.position.ReadValue() ?? default;
-
-        public static InputActionAsset actions => InputSystem.InputSystem.settings.actions;
-
         private static InputActionMap s_ButtonActions;
-        private static InputAction s_LeftMouseButtonAction;
-        private static InputAction s_RightMouseButtonAction;
-        private static InputAction s_MiddleMouseButtonAction;
-        private static InputAction s_ForwardMouseButtonAction;
-        private static InputAction s_BackMouseButtonAction;
         private static InputAction s_AnyKeyAction;
         private static KeySet s_PressedKeys;
         private static KeySet s_PressedKeysBefore;
@@ -293,66 +53,7 @@ namespace UnityEngine
         private static LocationService s_Location = new LocationService();
         private static Gyroscope s_Gyro = new Gyroscope();
         private static Compass s_Compass = new Compass();
-        private static int s_JoystickCount;
-        private static Joystick[] s_Joysticks;
         private static Action<InputDevice, InputDeviceChange> s_OnDeviceChangeDelegate;
-
-        private struct Joystick
-        {
-            public int index;
-            public string name;
-            public InputDevice device;
-            public InputActionMap map;
-
-            public bool connected => name != string.Empty;
-
-            public void Connect(int index, InputDevice device)
-            {
-                name = device.displayName;
-                this.device = device;
-                this.index = index;
-
-                // Tag the device such that bindings can bind to it by index.
-                // NOTE: Plus 1 here because joystick #0 is not an actual device but an aggregated
-                //       view of all joysticks.
-                InputSystem.InputSystem.AddDeviceUsage(device, "Joystick" + (index + 1));
-
-                if (map == null)
-                    map = CreateJoystickButtonMap();
-
-                map.devices = new[] { device };
-                map.Enable();
-            }
-
-            public void Disconnect()
-            {
-                name = string.Empty;
-                map.Disable();
-
-                // Remove tag.
-                InputSystem.InputSystem.RemoveDeviceUsage(device, "Joystick" + (index + 1));
-
-                // We keep the device around in case we reconnect.
-            }
-
-            public InputAction Button(int index)
-            {
-                return map.actions[index];
-            }
-
-            private static InputActionMap CreateJoystickButtonMap()
-            {
-                var map = new InputActionMap();
-                for (var i = 0; i < kMaxButtonsPerJoystickAsPerKeyCodeEnum; ++i)
-                {
-                    var code = KeyCode.JoystickButton0 + i;
-                    var action = map.AddAction("JoystickButton" + i, type: InputActionType.Button);
-                    foreach (var path in code.JoystickButtonToBindingPath())
-                        action.AddBinding(path);
-                }
-                return map;
-            }
-        }
 
         // We actually only support 5 ATM but keeping this at 7 for compatibility with legacy implementation.
         internal const int kMouseButtonCount = 7;
@@ -362,128 +63,6 @@ namespace UnityEngine
 
         // Same for buttons. We don't limit them but the KeyCode enum has a fixed set per joystick.
         internal const int kMaxButtonsPerJoystickAsPerKeyCodeEnum = 20;
-
-        private struct KeySet
-        {
-            public int Count => m_Count;
-            public Key this[int index] => (Key)m_Keys[index];
-
-            private int m_Count;
-            private byte[] m_Keys;
-
-            public void Add(Key key)
-            {
-                Debug.Assert(key != Key.None, "Must not try to enter Key.None into KeySet");
-                ArrayHelpers.AppendWithCapacity(ref m_Keys, ref m_Count, (byte)key);
-            }
-
-            public void Remove(Key key)
-            {
-                for (var i = 0; i < m_Count; ++i)
-                {
-                    if (m_Keys[i] == (byte)key)
-                    {
-                        ArrayHelpers.EraseAtByMovingTail(m_Keys, ref m_Count, i);
-                        return;
-                    }
-                }
-            }
-
-            public bool Contains(Key key)
-            {
-                for (var i = 0; i < m_Count; ++i)
-                    if (m_Keys[i] == (byte)key)
-                        return true;
-                return false;
-            }
-
-            public void Clear()
-            {
-                m_Count = 0;
-            }
-        }
-
-        private static unsafe void SyncKeys(Keyboard keyboard)
-        {
-            var statePtr = (KeyboardState*)keyboard.currentStatePtr;
-            MemoryHelpers.Swap(ref s_PressedKeys, ref s_PressedKeysBefore);
-            s_PressedKeys.Clear();
-
-            // Scan int by int.
-            for (var i = 0; i < KeyboardState.kSizeInInts; ++i)
-            {
-                var n = ((int*)statePtr->keys)[i];
-                if (n == 0)
-                    continue;
-
-                // Scan byte by byte.
-                for (var k = 0; k < 4; ++k)
-                {
-                    var b = (byte)(n & 0xff);
-                    n >>= 8;
-
-                    if (b == 0)
-                        continue;
-
-                    // Scan bit by bit.
-                    for (var j = 0; j < 8; ++j)
-                    {
-                        if ((b & (1 << j)) != 0)
-                        {
-                            var key = (Key)(i * 32 + k * 8 + j);
-                            s_PressedKeys.Add(key);
-                            if (!s_ThisFramePressedKeys.Contains(key))
-                                s_ThisFramePressedKeys.Add(key);
-                        }
-                    }
-                }
-            }
-
-            // Release keys that are no longer pressed.
-            for (var i = 0; i < s_PressedKeysBefore.Count; ++i)
-            {
-                var key = s_PressedKeysBefore[i];
-                if (!s_PressedKeys.Contains(key))
-                    s_ThisFrameReleasedKeys.Add(key);
-            }
-        }
-
-        private static void ReleaseKeys()
-        {
-            for (var i = 0; i < s_PressedKeys.Count; ++i)
-            {
-                var key = s_PressedKeys[i];
-                if (!s_ThisFrameReleasedKeys.Contains(key))
-                    s_ThisFrameReleasedKeys.Add(key);
-            }
-
-            s_PressedKeys.Clear();
-        }
-
-        private static InputAction GetActionForMouseButton(int button)
-        {
-            // Legacy input throws ArgumentException if button >= 7.
-            if (button < 0 || button >= kMouseButtonCount)
-                throw new ArgumentException("Invalid mouse button index.", nameof(button));
-
-            // Only support 3 buttons for now.
-            switch (button)
-            {
-                case 0:
-                    return s_LeftMouseButtonAction;
-                case 1:
-                    return s_RightMouseButtonAction;
-                case 2:
-                    return s_MiddleMouseButtonAction;
-                ////TODO: Old input Win32 backend maps 3=back(XBUTTON1) and 4=forward(XBUTTON2); find out what other platforms do
-                case 3:
-                    return s_BackMouseButtonAction;
-                case 4:
-                    return s_ForwardMouseButtonAction;
-            }
-
-            return null;
-        }
 
         internal static void OnEnteringPlayMode()
         {
@@ -564,49 +143,16 @@ namespace UnityEngine
 
                         if (change == InputDeviceChange.Added || change == InputDeviceChange.Reconnected)
                         {
-                            // Try to find the device in our s_Joysticks list.
-                            var index = GetJoystickIndex(device);
-                            if (index == -1)
-                            {
-                                // Not found. See if there is an empty slot in the joystick list.
-                                for (index = 0; index < s_JoystickCount; ++index)
-                                {
-                                    if (!s_Joysticks[index].connected)
-                                        break;
-                                }
-                            }
-
-                            if (index < 0 || index >= s_JoystickCount)
-                            {
-                                // No available entry. Append new joystick to end of array.
-                                index = ArrayHelpers.AppendWithCapacity(ref s_Joysticks, ref s_JoystickCount, default);
-                            }
-
-                            // Switch slot to joystick. Maybe a new connect, a reconnect, or a switch
-                            // to a different joystick/gamepad.
-                            s_Joysticks[index].Connect(index, device);
+                            AddJoystick(device);
                         }
                         else
                         {
-                            var index = GetJoystickIndex(device);
-                            if (index != -1)
-                            {
-                                // Mark joystick as disconnected.
-                                s_Joysticks[index].Disconnect();
-                            }
+                            RemoveJoystick(device);
                         }
                     }
                     break;
                 }
             }
-        }
-
-        private static int GetJoystickIndex(InputDevice device)
-        {
-            for (var i = 0; i < s_JoystickCount; ++i)
-                if (s_Joysticks[i].device == device)
-                    return i;
-            return -1;
         }
 
         internal static void NextFrame()
@@ -641,284 +187,6 @@ namespace UnityEngine
             InputSystem.InputSystem.onDeviceChange -= s_OnDeviceChangeDelegate;
         }
 
-        // We throw the same exceptions as the current native code.
-        [DoesNotReturn]
-        private static void AxisNotFound(string axisName)
-        {
-            throw new ArgumentException($"No input axis called {axisName} is defined; to set up input actions, go to: Edit -> Project Settings... -> Input", nameof(axisName));
-        }
-
-        private static KeyCode StringToKeyCode(string key)
-        {
-            if (key == null)
-                throw new ArgumentNullException(nameof(key));
-            var keyCode = key.ToKeyCode();
-            if (keyCode == null)
-                throw new ArgumentException($"Input Key named: {key} is unknown");
-            return keyCode.Value;
-        }
-
-        public static float GetAxis(string axisName)
-        {
-            var action = axisName != null ? actions.FindAction(axisName) : null;
-            if (action == null)
-            {
-                // We didn't find the action. See if the user is looking for "Vertical" and "Horizontal".
-                // If so, we alternatively look for a "Move" actions and take its X or Y value if found.
-
-                var isVertical = string.Equals(axisName, "Vertical", StringComparison.OrdinalIgnoreCase);
-                var isHorizontal = !isVertical && string.Equals(axisName, "Horizontal", StringComparison.OrdinalIgnoreCase);
-
-                if (isVertical || isHorizontal)
-                {
-                    var moveAction = actions.FindAction("Move");
-                    if (moveAction != null)
-                    {
-                        if (isVertical)
-                            return moveAction.ReadValue<Vector2>().y;
-                        return moveAction.ReadValue<Vector2>().x;
-                    }
-                }
-                else
-                {
-                    var isMouseX = string.Equals(axisName, "Mouse X", StringComparison.OrdinalIgnoreCase);
-                    var isMouseY = !isMouseX && string.Equals(axisName, "Mouse Y", StringComparison.OrdinalIgnoreCase);
-
-                    if (isMouseX || isMouseY)
-                    {
-                        var lookAction = actions.FindAction("Look");
-                        if (lookAction != null)
-                        {
-                            if (isMouseY)
-                                return lookAction.ReadValue<Vector2>().y;
-                            return lookAction.ReadValue<Vector2>().x;
-                        }
-                    }
-                }
-
-                AxisNotFound(axisName);
-            }
-
-            return action.ReadValue<float>();
-        }
-
-        private static InputAction GetButtonAction(string buttonName)
-        {
-            var action = buttonName != null ? actions.FindAction(buttonName) : null;
-            if (action == null)
-            {
-                if (buttonName == "Fire1")
-                    action = actions.FindAction("Fire");
-
-                if (action == null)
-                    AxisNotFound(buttonName);
-            }
-
-            return action;
-        }
-
-        public static bool GetButton(string buttonName)
-        {
-            return GetButtonAction(buttonName).IsPressed();
-        }
-
-        public static bool GetButtonUp(string buttonName)
-        {
-            return GetButtonAction(buttonName).WasReleasedThisFrame();
-        }
-
-        public static bool GetButtonDown(string buttonName)
-        {
-            return GetButtonAction(buttonName).WasPressedThisFrame();
-        }
-
-        public static bool GetMouseButton(int button)
-        {
-            var action = GetActionForMouseButton(button);
-            if (action == null)
-                return false;
-
-            return action.IsPressed();
-        }
-
-        public static bool GetMouseButtonUp(int button)
-        {
-            var action = GetActionForMouseButton(button);
-            if (action == null)
-                return false;
-
-            return action.WasReleasedThisFrame();
-        }
-
-        public static bool GetMouseButtonDown(int button)
-        {
-            var action = GetActionForMouseButton(button);
-            if (action == null)
-                return false;
-
-            return action.WasPressedThisFrame();
-        }
-
-        public static bool GetKey(string key)
-        {
-            return GetKey(StringToKeyCode(key));
-        }
-
-        public static bool GetKey(KeyCode key)
-        {
-            if (key.IsMouseButton())
-            {
-                var button = key - KeyCode.Mouse0;
-                return GetMouseButton(button);
-            }
-
-            if (key.IsJoystickButton())
-            {
-                var joystickIndex = key.GetJoystickIndex();
-                var buttonIndex = key.GetJoystickButtonIndex();
-
-                if (joystickIndex == 0)
-                {
-                    for (var i = 0; i < s_JoystickCount; ++i)
-                        if (s_Joysticks[i].Button(buttonIndex).IsPressed())
-                            return true;
-                    return false;
-                }
-
-                if (joystickIndex - 1 >= s_JoystickCount)
-                    return false;
-
-                return s_Joysticks[joystickIndex - 1].Button(buttonIndex).IsPressed();
-            }
-
-            var keyId = key.ToKey();
-            if (keyId == null)
-                return false;
-
-            return s_PressedKeys.Contains(keyId.Value);
-        }
-
-        public static bool GetKeyDown(string key)
-        {
-            return GetKeyDown(StringToKeyCode(key));
-        }
-
-        public static bool GetKeyDown(KeyCode key)
-        {
-            if (key.IsMouseButton())
-            {
-                var button = key - KeyCode.Mouse0;
-                return GetMouseButtonDown(button);
-            }
-
-            if (key.IsJoystickButton())
-            {
-                var joystickIndex = key.GetJoystickIndex();
-                var buttonIndex = key.GetJoystickButtonIndex();
-
-                if (joystickIndex == 0)
-                {
-                    var anyWasPressed = false;
-                    var noneWasAlreadyPressed = true;
-
-                    for (var i = 0; i < s_JoystickCount; ++i)
-                    {
-                        var button = s_Joysticks[i].Button(buttonIndex);
-                        var wasPressedThisFrame = button.WasPressedThisFrame();
-                        anyWasPressed |= wasPressedThisFrame;
-                        if (!wasPressedThisFrame)
-                            noneWasAlreadyPressed &= !button.IsPressed();
-                    }
-
-                    return anyWasPressed && noneWasAlreadyPressed;
-                }
-
-                if (joystickIndex - 1 >= s_JoystickCount)
-                    return false;
-
-                return s_Joysticks[joystickIndex - 1].Button(buttonIndex).WasPressedThisFrame();
-            }
-
-            var keyId = key.ToKey();
-            if (keyId == null)
-                return false;
-
-            return s_ThisFramePressedKeys.Contains(keyId.Value);
-        }
-
-        public static bool GetKeyUp(string key)
-        {
-            return GetKeyUp(StringToKeyCode(key));
-        }
-
-        public static bool GetKeyUp(KeyCode key)
-        {
-            if (key.IsMouseButton())
-            {
-                var button = key - KeyCode.Mouse0;
-                return GetMouseButtonUp(button);
-            }
-
-            if (key.IsJoystickButton())
-            {
-                var joystickIndex = key.GetJoystickIndex();
-                var buttonIndex = key.GetJoystickButtonIndex();
-
-                if (joystickIndex == 0)
-                {
-                    var anyWasReleased = false;
-                    var noneIsPressed = true;
-
-                    for (var i = 0; i < s_JoystickCount; ++i)
-                    {
-                        var button = s_Joysticks[i].Button(buttonIndex);
-                        var wasReleasedThisFrame = button.WasReleasedThisFrame();
-                        anyWasReleased |= wasReleasedThisFrame;
-                        if (!wasReleasedThisFrame)
-                            noneIsPressed &= !button.IsPressed();
-                    }
-
-                    return anyWasReleased && noneIsPressed;
-                }
-
-                if (joystickIndex - 1 >= s_JoystickCount)
-                    return false;
-
-                return s_Joysticks[joystickIndex - 1].Button(buttonIndex).WasReleasedThisFrame();
-            }
-
-            var keyId = key.ToKey();
-            if (keyId == null)
-                return false;
-
-            return s_ThisFrameReleasedKeys.Contains(keyId.Value);
-        }
-
-        public static Gyroscope gyro => s_Gyro;
-        public static bool isGyroAvailable => InputSystem.Gyroscope.current != null;
-        public static LocationService location => s_Location;
-        public static Compass compass => s_Compass;
-        public static Vector3 acceleration => InputRuntime.s_Instance.acceleration;
-
-        public static AccelerationEvent GetAccelerationEvent(int index)
-        {
-            return InputRuntime.s_Instance.GetAccelerationEvent(index);
-        }
-
-        public static int accelerationEventCount => InputRuntime.s_Instance.accelerationEventCount;
-
-        public static AccelerationEvent[] accelerationEvents
-        {
-            get
-            {
-                var count = accelerationEventCount;
-                var result = new AccelerationEvent[count];
-                for (var i = 0; i < count; ++i)
-                    result[i] = GetAccelerationEvent(i);
-                return result;
-            }
-        }
-
         #region Unimplemented
 
         public static float GetAxisRaw(string axisName)
@@ -938,36 +206,6 @@ namespace UnityEngine
 
         private static void SimulateTouch(Touch touch)
         {
-        }
-
-        public static string[] GetJoystickNames()
-        {
-            var result = new string[s_JoystickCount];
-            for (var i = 0; i < s_JoystickCount; ++i)
-                result[i] = s_Joysticks[i].name;
-            return result;
-        }
-
-        /*
-        public static int GetJoystickCount()
-        {
-            return m_JoystickNameCount;
-        }
-
-        public static string GetJoystickName(int index)
-        {
-            //...
-        }
-
-        public static InputDevice GetJoystick(int index)
-        {
-            //...
-        }
-        */
-
-        public static bool IsJoystickPreconfigured(string joystickName)
-        {
-            return false;
         }
 
         public static PenData GetPenEvent(int index)
