+++ Packages/com.unity.input/Runtime/Input_Axis.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 50875d4ef5ee4bec941558b04bbcfc8f
+timeCreated: 1649163690
