+++ Packages/com.unity.input/Runtime/Input_Joystick.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: b8bc004179774769b3642c6b93cc1b0e
+timeCreated: 1649163813
