+++ Packages/com.unity.input/Runtime/Input_Mouse.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 83d3ff63c9ab4dedac06d02b0977700e
+timeCreated: 1649163830
