+++ Packages/com.unity.input/Runtime/Input_Sensors.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9e0275f8f4e94b29aeab89eb086f80fe
+timeCreated: 1649172495
